Extracted sendResetNotification into auth.passwordreset
